refactor: Lifting creds_key/encryption_key into shared base settings by kent's m3
