+++ workflow-worker/tests/unit/test_config.py
@@ -14,7 +14,8 @@ def test_worker_settings_build_shared_client_configs() -> None:
         redis_uri="redis://redis:6379/1",
         redis_key_prefix="worker-prefix",
         jwt_private_key="test-private-key",
-        jwt_issuer="https://issuer.example.com",
+        # jwt_issuer is computed from auth_server_external_url (inherited), not set directly.
+        auth_server_external_url="https://issuer.example.com",
         jwt_self_signed_kid="test-kid",
         jwt_audience="test-audience",
         registry_app_name="test-worker",
@@ -28,6 +29,7 @@ def test_worker_settings_build_shared_client_configs() -> None:
     assert settings.mongo_config.mongodb_username == "worker"
     assert settings.redis_config.redis_uri == "redis://redis:6379/1"
     assert settings.redis_config.redis_key_prefix == "worker-prefix"
+    # jwt_issuer derives from auth_server_external_url → canonical scheme://netloc, matching registry.
     assert settings.jwt_signing_config.jwt_issuer == "https://issuer.example.com"
     assert settings.jwt_signing_config.registry_app_name == "test-worker"
     assert settings.encryption_key == bytes.fromhex("00" * 16)
@@ -41,13 +43,3 @@ def test_worker_settings_reject_claim_capacity_below_execution_capacity() -> Non
             creds_key="00" * 16,
             _env_file=None,
         )
-
-
-def test_worker_settings_reject_missing_creds_key() -> None:
-    with pytest.raises(ValidationError, match="CREDS_KEY must be set"):
-        WorkerSettings(creds_key="", _env_file=None)
-
-
-def test_worker_settings_reject_non_hex_creds_key() -> None:
-    with pytest.raises(ValidationError, match="CREDS_KEY must be a valid hex string"):
-        WorkerSettings(creds_key="not-hex", _env_file=None)
